% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_prob.R
\name{aipw_prob}
\alias{aipw_prob}
\title{Compute AIPW of survival probability}
\usage{
aipw_prob(data, tau)
}
\arguments{
\item{data}{A long form data frame (e.g., the output of transformData*())}

\item{tau}{A time of interest}
}
\value{
survival probability estimate for each treatment arm with influence function based standard error for the difference
}
\description{
Compute AIPW of survival probability
}
\examples{
# See example of tmle() for the creation of dlong
aipw_prob(dlong, tau)
}
