% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_rmst.R
\name{aipw_rmst}
\alias{aipw_rmst}
\title{Compute AIPW of RMST}
\usage{
aipw_rmst(data, tau)
}
\arguments{
\item{data}{A long form data frame (e.g., the output of transformData*())}

\item{tau}{A restriction time of interest}
}
\value{
RMST estimate for each treatment arm with influence function based standard error for the difference
}
\description{
Compute AIPW of RMST
}
\examples{
# See example of tmle() for the creation of dlong
aipw_rmst(dlong, tau)
}
