% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bound}
\alias{bound}
\title{Bound a vector of probabilities in [r, 1] for r > 0}
\usage{
bound(x, r = 0.001)
}
\arguments{
\item{x}{Vector of probabilities}

\item{r}{Bound}
}
\value{
Bounded vector of probabilities
}
\description{
Bound a vector of probabilities in [r, 1] for r > 0
}
