% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bound01}
\alias{bound01}
\title{Bound a vector of probabilities in [r, 1 - r] for r > 0}
\usage{
bound01(x, r = 1e-10)
}
\arguments{
\item{x}{Vector of probabilities}

\item{r}{Bound}
}
\value{
Bounded vector of probabilities
}
\description{
Bound a vector of probabilities in [r, 1 - r] for r > 0
}
