% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{ipw}
\alias{ipw}
\title{Compute IPW of RMST}
\usage{
ipw(data, tau)
}
\arguments{
\item{data}{A long form data frame (e.g., the output of transformData*())}

\item{tau}{A restriction time of interest}
}
\value{
RMST estimate for each treatment arm
}
\description{
Compute IPW of RMST
}
\examples{
# See example of tmle() for the creation of dlong
ipw(dlong, tau)
}
