% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_rmst.R
\name{ipw_rmst}
\alias{ipw_rmst}
\title{Compute IPW of RMST}
\usage{
ipw_rmst(data, tau)
}
\arguments{
\item{data}{A long form data frame (e.g., the output of transformData*())}

\item{tau}{A restriction time of interest}
}
\value{
RMST estimate for each treatment arm
}
\description{
Compute IPW of RMST
}
\examples{
# See example of tmle_rmst() for the creation of dlong
ipw_rmst(dlong, tau)
}
