% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{transformData}
\alias{transformData}
\title{Transform a survival dataset from short to long form}
\usage{
transformData(df, freq.time)
}
\arguments{
\item{df}{A data frame containing columns T (time to event) and C (censoring indicator)}

\item{freq.time}{If time T must be ccategorized into coarser intervals (e.g., use freq.time = 7 if T is in days and you want the output in weeks)}
}
\value{
A long form data set
}
\description{
Transform a survival dataset from short to long form
}
