% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_prob.R
\name{unadjusted_prob}
\alias{unadjusted_prob}
\title{Compute unadjusted Kaplan-Meier estimators of survival probability}
\usage{
unadjusted_prob(data, tau)
}
\arguments{
\item{data}{A long form data frame (e.g., the output of transformData*())}

\item{tau}{A time of interest}
}
\value{
survival probability estimate for each treatment arm
}
\description{
Compute unadjusted Kaplan-Meier estimators of survival probability
}
\examples{
# See example of tmle() for the creation of dlong
unadjusted(dlong, tau)
}
