% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_rmst.R
\name{unadjusted_rmst}
\alias{unadjusted_rmst}
\title{Compute unadjusted IPW and Kaplan-Meier estimators of RMST}
\usage{
unadjusted_rmst(data, tau)
}
\arguments{
\item{data}{A long form data frame (e.g., the output of transformData*())}

\item{tau}{A restriction time of interest}
}
\value{
RMST estimate for each treatment arm
}
\description{
Compute unadjusted IPW and Kaplan-Meier estimators of RMST
}
\examples{
# See example of tmle() for the creation of dlong
unadjusted(dlong, tau)
}
